main/requirements.txt
aniso8601
attrs
certifi
chardet
Click
Flask
flask-restx
Flask-Script
Flask-Testing
flask-cors
idna
itsdangerous
Jinja2
MarkupSafe
pycparser
pymongo
pyrsistent
python-simplexml
pytz
requests
six
urllib3
Werkzeug
xmltodict
jsonschema
responses
more-itertools
gunicorn
newrelic
marshmallow
virtualenv
ujson
Dumper
flask-dotenv
pyjwt
logger
Flask-Login
flask-blueprint
Flask-SQLAlchemy
SQLAlchemy
PyMySQL
pymongo
tzutil
dnspython

